Refactor SearchService
