+++ src/services/SearchService/index.ts
@@ -1,0 +1,3 @@
+export { SearchResult, SearchResultItem, LastResult } from './types';
+export { ScoringEngine } from './ScoringEngine';
+export { default } from './SearchService';
